backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
razorpay==1.4.1
python-telegram-bot==20.7
python-dotenv==1.0.0
bcrypt==4.1.2
